LPD-58417 Merge OpenIndexRequestExecutor/OpenIndexRequestExecutorImpl into pojos
