+++ modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/main/java/com/liferay/portal/search/elasticsearch7/internal/search/engine/adapter/index/ElasticsearchIndexRequestExecutor.java
@@ -157,6 +157,8 @@ public class ElasticsearchIndexRequestExecutor implements IndexRequestExecutor {
 			_elasticsearchClientResolver);
 		_getMappingIndexRequestExecutor = new GetMappingIndexRequestExecutor(
 			_elasticsearchClientResolver);
+		_openIndexRequestExecutor = new OpenIndexRequestExecutor(
+			_elasticsearchClientResolver);
 		_statsIndexRequestExecutor = new StatsIndexRequestExecutor(
 			_elasticsearchClientResolver, _jsonFactory);
 	}
@@ -194,7 +196,6 @@ public class ElasticsearchIndexRequestExecutor implements IndexRequestExecutor {
 	@Reference
 	private JSONFactory _jsonFactory;
 
-	@Reference
 	private OpenIndexRequestExecutor _openIndexRequestExecutor;
 
 	@Reference
