+++ modules/apps/portal-search-opensearch2/portal-search-opensearch2-impl/src/main/java/com/liferay/portal/search/opensearch2/internal/search/engine/adapter/index/OpenSearchIndexRequestExecutor.java
@@ -157,6 +157,8 @@ public class OpenSearchIndexRequestExecutor implements IndexRequestExecutor {
 			_openSearchConnectionManager);
 		_getMappingIndexRequestExecutor = new GetMappingIndexRequestExecutor(
 			_openSearchConnectionManager);
+		_openIndexRequestExecutor = new OpenIndexRequestExecutor(
+			_openSearchConnectionManager);
 		_statsIndexRequestExecutor = new StatsIndexRequestExecutor(
 			_openSearchConnectionManager);
 	}
@@ -188,7 +190,6 @@ public class OpenSearchIndexRequestExecutor implements IndexRequestExecutor {
 	private IndicesExistsIndexRequestExecutor
 		_indicesExistsIndexRequestExecutor;
 
-	@Reference
 	private OpenIndexRequestExecutor _openIndexRequestExecutor;
 
 	@Reference
