+++ modules/apps/portal-search-solr8/portal-search-solr8-impl/src/main/java/com/liferay/portal/search/solr8/internal/search/engine/adapter/index/OpenIndexRequestExecutor.java
@@ -11,8 +11,10 @@ import com.liferay.portal.search.engine.adapter.index.OpenIndexResponse;
 /**
  * @author Bryan Engler
  */
-public interface OpenIndexRequestExecutor {
+public class OpenIndexRequestExecutor {
 
-	public OpenIndexResponse execute(OpenIndexRequest openIndexRequest);
+	public OpenIndexResponse execute(OpenIndexRequest openIndexRequest) {
+		throw new UnsupportedOperationException();
+	}
 
 }
