--- modules/apps/portal-search-solr8/portal-search-solr8-impl/src/main/java/com/liferay/portal/search/solr8/internal/search/engine/adapter/index/OpenIndexRequestExecutorImpl.java
@@ -1,24 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.portal.search.solr8.internal.search.engine.adapter.index;
-
-import com.liferay.portal.search.engine.adapter.index.OpenIndexRequest;
-import com.liferay.portal.search.engine.adapter.index.OpenIndexResponse;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Bryan Engler
- */
-@Component(service = OpenIndexRequestExecutor.class)
-public class OpenIndexRequestExecutorImpl implements OpenIndexRequestExecutor {
-
-	@Override
-	public OpenIndexResponse execute(OpenIndexRequest openIndexRequest) {
-		throw new UnsupportedOperationException();
-	}
-
-}
