+++ modules/apps/portal-search-solr8/portal-search-solr8-impl/src/main/java/com/liferay/portal/search/solr8/internal/search/engine/adapter/index/SolrIndexRequestExecutor.java
@@ -176,8 +176,8 @@ public class SolrIndexRequestExecutor implements IndexRequestExecutor {
 	private IndicesExistsIndexRequestExecutor
 		_indicesExistsIndexRequestExecutor;
 
-	@Reference
-	private OpenIndexRequestExecutor _openIndexRequestExecutor;
+	private final OpenIndexRequestExecutor _openIndexRequestExecutor =
+		new OpenIndexRequestExecutor();
 
 	@Reference
 	private PutMappingIndexRequestExecutor _putMappingIndexRequestExecutor;
